Centralize delete of data set
