+++ timbuctoo-instancev4/src/main/java/nl/knaw/huygens/timbuctoo/v5/datastorage/DataSetStorage.java
@@ -23,4 +23,6 @@ public interface DataSetStorage {
   File getResourceSyncDescriptionFile();
 
   JsonDataStore<LogList> getLogList() throws IOException;
+
+  void clear() throws IOException;
 }
