+++ timbuctoo-instancev4/src/main/java/nl/knaw/huygens/timbuctoo/v5/filestorage/FileStorage.java
@@ -26,4 +26,6 @@ public interface FileStorage {
    */
   Optional<CachedFile> getFile(String token) throws IOException;
 
+  void clear() throws IOException;
+
 }
