+++ timbuctoo-instancev4/src/main/java/nl/knaw/huygens/timbuctoo/v5/jsonfilebackeddata/JsonFileBackedData.java
@@ -41,6 +41,12 @@ public class JsonFileBackedData<T> implements JsonDataStore<T> {
     }
   }
 
+  public static void remove(File file) throws IOException {
+    synchronized (existing) {
+      existing.remove(file.getCanonicalPath());
+    }
+  }
+
 
   private final File file;
   private T value;
